Reorganize code structure in modules
