--- src/models/category.ts
@@ -1,15 +1,0 @@
-import mongoose from 'mongoose';
-import { prettifyId } from '@/utils/dbUtils';
-
-const CategorySchema = new mongoose.Schema({
-  name: String,
-  code: { type: String, unique: true, required: true },
-  events: [{
-    type: mongoose.Schema.ObjectId,
-    ref: 'Event',
-  }]
-});
-prettifyId(CategorySchema);
-const Category = mongoose.model('Category', CategorySchema)
-
-export default Category;
