+++ src/modules/events/controllers/events.ts
@@ -1,0 +1,15 @@
+import express from 'express'
+import eventsService from '@/modules/events/services/events'
+
+const router = express.Router()
+
+router.get('/', async (req, res, next) => {
+  try {
+    const result = await eventsService.listEvents()
+    res.json(result).status(200)
+  } catch (err) {
+    next(err)
+  }
+})
+
+export default router
