+++ src/modules/events/repo/event.ts
@@ -1,0 +1,12 @@
+import mongoose from 'mongoose'
+import { prettifyId } from '@/infra/db/utils'
+
+const EventSchema = new mongoose.Schema({
+  name: String,
+  category_code: { type: String, required: true },
+  code: { type: String, required: true },
+})
+prettifyId(EventSchema)
+const Event = mongoose.model('Event', EventSchema)
+
+export default Event
