+++ src/modules/events/services/events.ts
@@ -1,0 +1,9 @@
+import Event from '@/modules/events/repo/event'
+
+const listEvents = async () => {
+  return await Event.find({})
+}
+
+export default {
+  listEvents,
+}
