--- src/services/categories.ts
@@ -1,12 +1,0 @@
-import Category from '@/models/category';
-
-const getCategories = async (populate) => {
-  if (populate && populate.toLowerCase() === 'true') {
-    return await Category.find({}).populate('events');
-  }
-  return await Category.find({});
-}
-
-export default {
-  getCategories
-}
