--- src/services/events.ts
@@ -1,9 +1,0 @@
-import Event from '@/models/event';
-
-const listEvents = async () => {
-  return await Event.find({});
-}
-
-export default {
-  listEvents
-}
