--- src/utils/for_testing.ts
@@ -1,19 +1,0 @@
-const reverse = (string) => {
-  return string
-    .split('')
-    .reverse()
-    .join('')
-}
-
-const average = (array) => {
-  const reducer = (sum, item) => {
-    return sum + item
-  }
-
-  return !array.length ? 0 : array.reduce(reducer, 0) / array.length;
-}
-
-export {
-  reverse,
-  average,
-}
